azizbek_xabibullayev/selenium/requirements.txt
selenium==4.31.0
psycopg2-binary==2.9.10
webdriver-manager==4.0.2